Fix a bug in extract_best_per_route kernel (#156)
The mentioned kernel does not require any dynamic shared memory, however, we are passing `sh_size` that is relevant for the previous kernel in the code.  In most cases it is fine. However, if the sh_size is more than the shared memory available, the kernel fails to launch with `cudaInvalidValue` error.   The previous kernel has no issues because the dynamic shared memory is tuned for that kernel using `cudaFuncSetAttribute` call.  

This PR passes zero as the dynamic shared size to resolve the issue. 

## Issue


This is a bug reported by a customer.

Authors:
  - Rajesh Gandham (https://github.com/rg20)

Approvers:
  - Hugo Linsenmaier (https://github.com/hlinsen)

URL: https://github.com/NVIDIA/cuopt/pull/156