Compute relative gap with respect to user objectives (#418)
In branch-and-bound, to stop the search, we are currently computing the relative gap based on the working problem.  When the objective offset is very large,  the relative gap on the user problem can be very large even when it is small on the working problem.  This PR fixes this by always computing the relative gap with respect to the user problem.  

## Issue

Closes #417

Authors:
  - Rajesh Gandham (https://github.com/rg20)

Approvers:
  - Alice Boucher (https://github.com/aliceb-nv)

URL: https://github.com/NVIDIA/cuopt/pull/418