Barrier post-solve, find_dense_columns fix, folding fix (#462)
Fixes a crash on physciansched6-2 in barrier post-solve.

Fixes a crash on roi5alpha10n8 in `find_dense_columns`. There was an undefined memory access because the number of nonzeros in the row was greater than the number of rows. This fix correctly sets the size of the array to the number of columns.

Fixes an undefined memory reference on roi5alpha10n8 in folding. The colors array was being realloced an so the reference to the refining color pointed to memory that had been freed. This fix gets a new reference to the refining color on each iteration of the loop that splits colors.

Authors:
  - Chris Maes (https://github.com/chris-maes)

Approvers:
  - Rajesh Gandham (https://github.com/rg20)
  - Hugo Linsenmaier (https://github.com/hlinsen)

URL: https://github.com/NVIDIA/cuopt/pull/462