Fix bug in logger when log_to_console is set to false (#538)
Prior to this PR,  some logging info relating to solver parameters was shown in stdout.  This is because we initialize the logger to stdout and we do not have apriori information whether the corresponding parameter would be set or not. In this PR, I am using the callback logger along with a log buffer to store all the logging info until the appropriate parameters and the loggers are initialized. 

## Issue

Closes #187

Authors:
  - Rajesh Gandham (https://github.com/rg20)

Approvers:
  - Alice Boucher (https://github.com/aliceb-nv)
  - Akif ÇÖRDÜK (https://github.com/akifcorduk)
  - Trevor McKay (https://github.com/tmckayus)

URL: https://github.com/NVIDIA/cuopt/pull/538